Refactor: Move robot configuration class, fix movement

+++ Code/UiRemote/src/RobotRemote/RobotServices/Movement/IMovementService.java
@@ -1,6 +1,6 @@
 package RobotRemote.RobotServices.Movement;
 
-import RobotRemote.Models.RobotConfiguration;
+import RobotRemote.Shared.RobotConfiguration;
 import RobotRemote.RobotServices.Connection.RobotConnectionService;
 import RobotRemote.Shared.AppStateRepository;
 
@@ -10,7 +10,7 @@ public interface IMovementService {
   void forward(float dist_cm);
   void backward();
   void backward(float dist_cm);
-  void turn(float degrees);
+  void turn(int degrees);
   boolean isMoving();
   void Initialize(RobotConfiguration configuration, RobotConnectionService robotConnectionService, AppStateRepository appStateRepository);
